# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {"title":"NSSMS API","version":"1.0.0","description":"Versioned API for the National School Sports Management System."},
  "servers":[{"url":"http://localhost:3000"}],
  "components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer"}}},
  "paths":{
...
    "/api/v1/auth/login":{"post":{"summary":"Authenticate an administrator","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["username","password"],"properties":{"username":{"type":"string"},"password":{"type":"string","format":"password"}}}}}},"responses":{"200":{"description":"Authenticated"},"401":{"description":"Invalid credentials"}}}},
    "/api/v1/auth/logout":{"post":{"security":[{"bearerAuth":[]}],"summary":"Record an administrative logout","responses":{"200":{"description":"Logged out"},"401":{"description":"Unauthorized"}}}},
    "/api/v1/auth/me":{"get":{"security":[{"bearerAuth":[]}],"summary":"Read the authenticated administrator","responses":{"200":{"description":"Authenticated user"},"401":{"description":"Unauthorized"}}}},
    "/api/v1/public/licenses/verify/{reference}":{"get":{"summary":"Publicly verify a license without exposing internal identifiers","parameters":[{"name":"reference","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Approved public verification projection"},"404":{"description":"Not found"}}}},
    "/api/v1/public/seasons":{"get":{"summary":"List approved public seasons","parameters":[{"name":"page","in":"query","schema":{"type":"integer","minimum":1}},{"name":"pageSize","in":"query","schema":{"type":"integer","minimum":1,"maximum":100}}],"responses":{"200":{"description":"Public season list"}}}},
    "/api/v1/public/competitions":{"get":{"summary":"List approved public competitions","parameters":[{"name":"page","in":"query","schema":{"type":"integer","minimum":1}},{"name":"pageSize","in":"query","schema":{"type":"integer","minimum":1,"maximum":100}}],"responses":{"200":{"description":"Public competition list"}}}},
    "/api/v1/public/results":{"get":{"summary":"List published results","parameters":[{"name":"page","in":"query","schema":{"type":"integer","minimum":1}},{"name":"pageSize","in":"query","schema":{"type":"integer","minimum":1,"maximum":100}}],"responses":{"200":{"description":"Public results list"}}}},
...
    "/api/v1/public/geography/wilayas/{id}/dairas":{"get":{"summary":"List dairas for a wilaya","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","minimum":1,"maximum":58}}],"responses":{"200":{"description":"Daira list"},"400":{"description":"Invalid wilaya id"}}}},
    "/api/v1/public/geography/dairas/{id}/communes":{"get":{"summary":"List communes for a daira","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","minimum":0}}],"responses":{"200":{"description":"Commune list"},"400":{"description":"Invalid daira id"}}}},
...
  }
}